[ML] refactor tasks classes and add new abstract executor
