+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/action/TransportOpenJobAction.java
@@ -11,70 +11,49 @@ import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.ElasticsearchStatusException;
 import org.elasticsearch.ResourceAlreadyExistsException;
-import org.elasticsearch.ResourceNotFoundException;
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.support.ActionFilters;
-import org.elasticsearch.action.support.IndicesOptions;
 import org.elasticsearch.action.support.master.TransportMasterNodeAction;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.block.ClusterBlockException;
 import org.elasticsearch.cluster.block.ClusterBlockLevel;
 import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
-import org.elasticsearch.cluster.node.DiscoveryNode;
-import org.elasticsearch.cluster.routing.IndexRoutingTable;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.TimeValue;
-import org.elasticsearch.index.engine.DocumentMissingException;
 import org.elasticsearch.license.LicenseUtils;
 import org.elasticsearch.license.XPackLicenseState;
-import org.elasticsearch.persistent.AllocatedPersistentTask;
-import org.elasticsearch.persistent.PersistentTaskState;
 import org.elasticsearch.persistent.PersistentTasksCustomMetadata;
-import org.elasticsearch.persistent.PersistentTasksExecutor;
 import org.elasticsearch.persistent.PersistentTasksService;
 import org.elasticsearch.persistent.decider.EnableAssignmentDecider;
 import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.tasks.Task;
-import org.elasticsearch.tasks.TaskId;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportService;
 import org.elasticsearch.xpack.core.XPackField;
 import org.elasticsearch.xpack.core.ml.MlConfigIndex;
-import org.elasticsearch.xpack.core.ml.MlMetaIndex;
-import org.elasticsearch.xpack.core.ml.MlMetadata;
 import org.elasticsearch.xpack.core.ml.MlTasks;
-import org.elasticsearch.xpack.core.ml.action.FinalizeJobExecutionAction;
 import org.elasticsearch.xpack.core.ml.action.NodeAcknowledgedResponse;
 import org.elasticsearch.xpack.core.ml.action.OpenJobAction;
 import org.elasticsearch.xpack.core.ml.job.config.Job;
 import org.elasticsearch.xpack.core.ml.job.config.JobState;
 import org.elasticsearch.xpack.core.ml.job.config.JobTaskState;
 import org.elasticsearch.xpack.core.ml.job.config.JobUpdate;
-import org.elasticsearch.xpack.core.ml.job.persistence.AnomalyDetectorsIndex;
 import org.elasticsearch.xpack.core.ml.job.persistence.ElasticsearchMappings;
 import org.elasticsearch.xpack.core.ml.utils.ExceptionsHelper;
-import org.elasticsearch.xpack.ml.MachineLearning;
 import org.elasticsearch.xpack.ml.MlConfigMigrationEligibilityCheck;
 import org.elasticsearch.xpack.ml.job.JobNodeSelector;
 import org.elasticsearch.xpack.ml.job.persistence.JobConfigProvider;
-import org.elasticsearch.xpack.ml.job.persistence.JobResultsProvider;
-import org.elasticsearch.xpack.ml.job.process.autodetect.AutodetectProcessManager;
 import org.elasticsearch.xpack.ml.process.MlMemoryTracker;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.function.Predicate;
 
-import static org.elasticsearch.xpack.core.ClientHelper.ML_ORIGIN;
-import static org.elasticsearch.xpack.core.ClientHelper.executeAsyncWithOrigin;
 import static org.elasticsearch.xpack.core.ml.MlTasks.AWAITING_UPGRADE;
-import static org.elasticsearch.xpack.ml.MachineLearning.MAX_OPEN_JOBS_PER_NODE;
-import static org.elasticsearch.xpack.ml.MachineLearning.USE_AUTO_MACHINE_MEMORY_PERCENT;
+import static org.elasticsearch.xpack.ml.job.task.OpenJobPersistentTasksExecutor.makeAssignmentsNotAllowedException;
+import static org.elasticsearch.xpack.ml.job.task.OpenJobPersistentTasksExecutor.makeCurrentlyBeingUpgradedException;
+import static org.elasticsearch.xpack.ml.job.task.OpenJobPersistentTasksExecutor.makeNoSuitableNodesException;
 
 /*
  This class extends from TransportMasterNodeAction for cluster state observing purposes.
@@ -88,9 +67,6 @@ import static org.elasticsearch.xpack.ml.MachineLearning.USE_AUTO_MACHINE_MEMORY
 public class TransportOpenJobAction extends TransportMasterNodeAction<OpenJobAction.Request, NodeAcknowledgedResponse> {
 
     private static final Logger logger = LogManager.getLogger(TransportOpenJobAction.class);
-
-    static final PersistentTasksCustomMetadata.Assignment AWAITING_MIGRATION =
-            new PersistentTasksCustomMetadata.Assignment(null, "job cannot be assigned until it has been migrated.");
 
     private final XPackLicenseState licenseState;
     private final PersistentTasksService persistentTasksService;
@@ -113,82 +89,6 @@ public class TransportOpenJobAction extends TransportMasterNodeAction<OpenJobAct
         this.memoryTracker = memoryTracker;
         this.migrationEligibilityCheck = new MlConfigMigrationEligibilityCheck(settings, clusterService);
         this.client = client;
-    }
-
-    /**
-     * Validations to fail fast before trying to update the job state on master node:
-     * <ul>
-     *     <li>check job exists</li>
-     *     <li>check job is not marked as deleted</li>
-     *     <li>check job's version is supported</li>
-     * </ul>
-     */
-    static void validate(String jobId, Job job) {
-        if (job == null) {
-            throw ExceptionsHelper.missingJobException(jobId);
-        }
-        if (job.isDeleting()) {
-            throw ExceptionsHelper.conflictStatusException("Cannot open job [" + jobId + "] because it is being deleted");
-        }
-        if (job.getJobVersion() == null) {
-            throw ExceptionsHelper.badRequestException("Cannot open job [" + jobId
-                    + "] because jobs created prior to version 5.5 are not supported");
-        }
-    }
-
-    static String[] indicesOfInterest(String resultsIndex) {
-        if (resultsIndex == null) {
-            return new String[]{AnomalyDetectorsIndex.jobStateIndexPattern(), MlMetaIndex.indexName(),
-                MlConfigIndex.indexName()};
-        }
-        return new String[]{AnomalyDetectorsIndex.jobStateIndexPattern(), resultsIndex, MlMetaIndex.indexName(),
-            MlConfigIndex.indexName()};
-    }
-
-    static List<String> verifyIndicesPrimaryShardsAreActive(String resultsWriteIndex, ClusterState clusterState,
-                                                            IndexNameExpressionResolver expressionResolver) {
-        String[] indices = expressionResolver.concreteIndexNames(clusterState, IndicesOptions.lenientExpandOpen(),
-            indicesOfInterest(resultsWriteIndex));
-        List<String> unavailableIndices = new ArrayList<>(indices.length);
-        for (String index : indices) {
-            // Indices are created on demand from templates.
-            // It is not an error if the index doesn't exist yet
-            if (clusterState.metadata().hasIndex(index) == false) {
-                continue;
-            }
-            IndexRoutingTable routingTable = clusterState.getRoutingTable().index(index);
-            if (routingTable == null || routingTable.allPrimaryShardsActive() == false) {
-                unavailableIndices.add(index);
-            }
-        }
-        return unavailableIndices;
-    }
-
-    private static boolean nodeSupportsModelSnapshotVersion(DiscoveryNode node, Job job) {
-        if (job.getModelSnapshotId() == null || job.getModelSnapshotMinVersion() == null) {
-            // There is no snapshot to restore or the min model snapshot version is 5.5.0
-            // which is OK as we have already checked the node is >= 5.5.0.
-            return true;
-        }
-        return node.getVersion().onOrAfter(job.getModelSnapshotMinVersion());
-    }
-
-    public static String nodeFilter(DiscoveryNode node, Job job) {
-
-        String jobId = job.getId();
-
-        if (TransportOpenJobAction.nodeSupportsModelSnapshotVersion(node, job) == false) {
-            return "Not opening job [" + jobId + "] on node [" + JobNodeSelector.nodeNameAndVersion(node)
-                + "], because the job's model snapshot requires a node of version ["
-                + job.getModelSnapshotMinVersion() + "] or higher";
-        }
-
-        if (Job.getCompatibleJobTypes(node.getVersion()).contains(job.getJobType()) == false) {
-            return "Not opening job [" + jobId + "] on node [" + JobNodeSelector.nodeNameAndVersion(node) +
-                "], because this node does not support jobs of type [" + job.getJobType() + "]";
-        }
-
-        return null;
     }
 
     @Override
@@ -349,222 +249,6 @@ public class TransportOpenJobAction extends TransportMasterNodeAction<OpenJobAct
         );
     }
 
-    public static class OpenJobPersistentTasksExecutor extends PersistentTasksExecutor<OpenJobAction.JobParams> {
-
-        private static final Logger logger = LogManager.getLogger(OpenJobPersistentTasksExecutor.class);
-
-        private final AutodetectProcessManager autodetectProcessManager;
-        private final MlMemoryTracker memoryTracker;
-        private final Client client;
-        private final IndexNameExpressionResolver expressionResolver;
-        private final JobResultsProvider jobResultsProvider;
-        private final boolean useAutoMemoryPercentage;
-
-        private volatile int maxConcurrentJobAllocations;
-        private volatile int maxMachineMemoryPercent;
-        private volatile int maxLazyMLNodes;
-        private volatile int maxOpenJobs;
-        private volatile ClusterState clusterState;
-
-        public OpenJobPersistentTasksExecutor(Settings settings, ClusterService clusterService,
-                                              AutodetectProcessManager autodetectProcessManager, MlMemoryTracker memoryTracker,
-                                              Client client, IndexNameExpressionResolver expressionResolver) {
-            super(MlTasks.JOB_TASK_NAME, MachineLearning.UTILITY_THREAD_POOL_NAME);
-            this.autodetectProcessManager = Objects.requireNonNull(autodetectProcessManager);
-            this.memoryTracker = Objects.requireNonNull(memoryTracker);
-            this.client = Objects.requireNonNull(client);
-            this.expressionResolver = Objects.requireNonNull(expressionResolver);
-            this.jobResultsProvider = new JobResultsProvider(client, settings, expressionResolver);
-            this.maxConcurrentJobAllocations = MachineLearning.CONCURRENT_JOB_ALLOCATIONS.get(settings);
-            this.maxMachineMemoryPercent = MachineLearning.MAX_MACHINE_MEMORY_PERCENT.get(settings);
-            this.maxLazyMLNodes = MachineLearning.MAX_LAZY_ML_NODES.get(settings);
-            this.maxOpenJobs = MAX_OPEN_JOBS_PER_NODE.get(settings);
-            this.useAutoMemoryPercentage = USE_AUTO_MACHINE_MEMORY_PERCENT.get(settings);
-            clusterService.getClusterSettings()
-                    .addSettingsUpdateConsumer(MachineLearning.CONCURRENT_JOB_ALLOCATIONS, this::setMaxConcurrentJobAllocations);
-            clusterService.getClusterSettings()
-                    .addSettingsUpdateConsumer(MachineLearning.MAX_MACHINE_MEMORY_PERCENT, this::setMaxMachineMemoryPercent);
-            clusterService.getClusterSettings().addSettingsUpdateConsumer(MachineLearning.MAX_LAZY_ML_NODES, this::setMaxLazyMLNodes);
-            clusterService.getClusterSettings().addSettingsUpdateConsumer(MAX_OPEN_JOBS_PER_NODE, this::setMaxOpenJobs);
-            clusterService.addListener(event -> clusterState = event.state());
-        }
-
-        @Override
-        public PersistentTasksCustomMetadata.Assignment getAssignment(OpenJobAction.JobParams params, ClusterState clusterState) {
-
-            // If the task parameters do not have a job field then the job
-            // was first opened on a pre v6.6 node and has not been migrated
-            Job job = params.getJob();
-            if (job == null) {
-                return AWAITING_MIGRATION;
-            }
-
-            // If we are waiting for an upgrade to complete, we should not assign to a node
-            if (MlMetadata.getMlMetadata(clusterState).isUpgradeMode()) {
-                return AWAITING_UPGRADE;
-            }
-
-            String jobId = params.getJobId();
-            String resultsWriteAlias = AnomalyDetectorsIndex.resultsWriteAlias(jobId);
-            List<String> unavailableIndices = verifyIndicesPrimaryShardsAreActive(resultsWriteAlias, clusterState, expressionResolver);
-            if (unavailableIndices.size() != 0) {
-                String reason = "Not opening job [" + jobId + "], because not all primary shards are active for the following indices [" +
-                    String.join(",", unavailableIndices) + "]";
-                logger.debug(reason);
-                return new PersistentTasksCustomMetadata.Assignment(null, reason);
-            }
-
-            boolean isMemoryTrackerRecentlyRefreshed = memoryTracker.isRecentlyRefreshed();
-            if (isMemoryTrackerRecentlyRefreshed == false) {
-                boolean scheduledRefresh = memoryTracker.asyncRefresh();
-                if (scheduledRefresh) {
-                    String reason = "Not opening job [" + jobId + "] because job memory requirements are stale - refresh requested";
-                    logger.debug(reason);
-                    return new PersistentTasksCustomMetadata.Assignment(null, reason);
-                }
-            }
-
-            JobNodeSelector jobNodeSelector = new JobNodeSelector(clusterState, jobId, MlTasks.JOB_TASK_NAME, memoryTracker,
-                job.allowLazyOpen() ? Integer.MAX_VALUE : maxLazyMLNodes, node -> nodeFilter(node, job));
-            return jobNodeSelector.selectNode(
-                maxOpenJobs,
-                maxConcurrentJobAllocations,
-                maxMachineMemoryPercent,
-                isMemoryTrackerRecentlyRefreshed,
-                useAutoMemoryPercentage);
-        }
-
-        @Override
-        public void validate(OpenJobAction.JobParams params, ClusterState clusterState) {
-
-            TransportOpenJobAction.validate(params.getJobId(), params.getJob());
-
-            // If we already know that we can't find an ml node because all ml nodes are running at capacity or
-            // simply because there are no ml nodes in the cluster then we fail quickly here:
-            PersistentTasksCustomMetadata.Assignment assignment = getAssignment(params, clusterState);
-            if (assignment.equals(AWAITING_UPGRADE)) {
-                throw makeCurrentlyBeingUpgradedException(logger, params.getJobId(), assignment.getExplanation());
-            }
-
-            if (assignment.getExecutorNode() == null && assignment.equals(JobNodeSelector.AWAITING_LAZY_ASSIGNMENT) == false) {
-                throw makeNoSuitableNodesException(logger, params.getJobId(), assignment.getExplanation());
-            }
-        }
-
-        @Override
-        protected void nodeOperation(AllocatedPersistentTask task, OpenJobAction.JobParams params, PersistentTaskState state) {
-            JobTask jobTask = (JobTask) task;
-            jobTask.autodetectProcessManager = autodetectProcessManager;
-            JobTaskState jobTaskState = (JobTaskState) state;
-            JobState jobState = jobTaskState == null ? null : jobTaskState.getState();
-            jobResultsProvider.setRunningForecastsToFailed(params.getJobId(), ActionListener.wrap(
-                r -> runJob(jobTask, jobState, params),
-                e -> {
-                    logger.warn(new ParameterizedMessage("[{}] failed to set forecasts to failed", params.getJobId()), e);
-                    runJob(jobTask, jobState, params);
-                }
-            ));
-        }
-
-        private void runJob(JobTask jobTask, JobState jobState, OpenJobAction.JobParams params) {
-            // If the job is closing, simply stop and return
-            if (JobState.CLOSING.equals(jobState)) {
-                // Mark as completed instead of using `stop` as stop assumes native processes have started
-                logger.info("[{}] job got reassigned while stopping. Marking as completed", params.getJobId());
-                jobTask.markAsCompleted();
-                return;
-            }
-            // If the job is failed then the Persistent Task Service will
-            // try to restart it on a node restart. Exiting here leaves the
-            // job in the failed state and it must be force closed.
-            if (JobState.FAILED.equals(jobState)) {
-                return;
-            }
-
-            String jobId = jobTask.getJobId();
-            autodetectProcessManager.openJob(jobTask, clusterState, (e2, shouldFinalizeJob) -> {
-                if (e2 == null) {
-                    if (shouldFinalizeJob) {
-                        FinalizeJobExecutionAction.Request finalizeRequest = new FinalizeJobExecutionAction.Request(new String[]{jobId});
-                        executeAsyncWithOrigin(client, ML_ORIGIN, FinalizeJobExecutionAction.INSTANCE, finalizeRequest,
-                            ActionListener.wrap(
-                                response -> jobTask.markAsCompleted(),
-                                e -> {
-                                    logger.error("error finalizing job [" + jobId + "]", e);
-                                    Throwable unwrapped = ExceptionsHelper.unwrapCause(e);
-                                    if (unwrapped instanceof DocumentMissingException || unwrapped instanceof ResourceNotFoundException) {
-                                        jobTask.markAsCompleted();
-                                    } else {
-                                        jobTask.markAsFailed(e);
-                                    }
-                                }
-                            ));
-                    } else {
-                        jobTask.markAsCompleted();
-                    }
-                } else {
-                    jobTask.markAsFailed(e2);
-                }
-            });
-        }
-
-        @Override
-        protected AllocatedPersistentTask createTask(long id, String type, String action, TaskId parentTaskId,
-                                                     PersistentTasksCustomMetadata.PersistentTask<OpenJobAction.JobParams> persistentTask,
-                                                     Map<String, String> headers) {
-            return new JobTask(persistentTask.getParams().getJobId(), id, type, action, parentTaskId, headers);
-        }
-
-        void setMaxConcurrentJobAllocations(int maxConcurrentJobAllocations) {
-            this.maxConcurrentJobAllocations = maxConcurrentJobAllocations;
-        }
-
-        void setMaxMachineMemoryPercent(int maxMachineMemoryPercent) {
-            this.maxMachineMemoryPercent = maxMachineMemoryPercent;
-        }
-
-        void setMaxLazyMLNodes(int maxLazyMLNodes) {
-            this.maxLazyMLNodes = maxLazyMLNodes;
-        }
-
-        void setMaxOpenJobs(int maxOpenJobs) {
-            this.maxOpenJobs = maxOpenJobs;
-        }
-
-    }
-
-    public static class JobTask extends AllocatedPersistentTask implements OpenJobAction.JobTaskMatcher {
-
-        private static final Logger LOGGER = LogManager.getLogger(JobTask.class);
-
-        private final String jobId;
-        private volatile AutodetectProcessManager autodetectProcessManager;
-
-        JobTask(String jobId, long id, String type, String action, TaskId parentTask, Map<String, String> headers) {
-            super(id, type, action, "job-" + jobId, parentTask, headers);
-            this.jobId = jobId;
-        }
-
-        public String getJobId() {
-            return jobId;
-        }
-
-        @Override
-        protected void onCancelled() {
-            String reason = getReasonCancelled();
-            LOGGER.trace("[{}] Cancelling job task because: {}", jobId, reason);
-            killJob(reason);
-        }
-
-        void killJob(String reason) {
-            autodetectProcessManager.killProcess(this, false, reason);
-        }
-
-        void closeJob(String reason) {
-            autodetectProcessManager.closeJob(this, false, reason);
-        }
-    }
-
     /**
      * This class contains the wait logic for waiting for a job's persistent task to be allocated on
      * job opening.  It should only be used in the open job action, and never at other times the job's
@@ -635,26 +319,5 @@ public class TransportOpenJobAction extends TransportMasterNodeAction<OpenJobAct
                     return true;
             }
         }
-    }
-
-    static ElasticsearchException makeNoSuitableNodesException(Logger logger, String jobId, String explanation) {
-        String msg = "Could not open job because no suitable nodes were found, allocation explanation [" + explanation + "]";
-        logger.warn("[{}] {}", jobId, msg);
-        Exception detail = new IllegalStateException(msg);
-        return new ElasticsearchStatusException("Could not open job because no ML nodes with sufficient capacity were found",
-            RestStatus.TOO_MANY_REQUESTS, detail);
-    }
-
-    static ElasticsearchException makeAssignmentsNotAllowedException(Logger logger, String jobId) {
-        String msg = "Cannot open jobs because persistent task assignment is disabled by the ["
-            + EnableAssignmentDecider.CLUSTER_TASKS_ALLOCATION_ENABLE_SETTING.getKey() + "] setting";
-        logger.warn("[{}] {}", jobId, msg);
-        return new ElasticsearchStatusException(msg, RestStatus.TOO_MANY_REQUESTS);
-    }
-
-    static ElasticsearchException makeCurrentlyBeingUpgradedException(Logger logger, String jobId, String explanation) {
-        String msg = "Cannot open jobs when upgrade mode is enabled";
-        logger.warn("[{}] {}", jobId, msg);
-        return new ElasticsearchStatusException(msg, RestStatus.TOO_MANY_REQUESTS);
     }
 }
