+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/action/TransportPostDataAction.java
@@ -16,6 +16,7 @@ import org.elasticsearch.xpack.core.ml.action.PostDataAction;
 import org.elasticsearch.xpack.ml.job.process.autodetect.AutodetectProcessManager;
 import org.elasticsearch.xpack.ml.job.process.autodetect.params.DataLoadParams;
 import org.elasticsearch.xpack.ml.job.process.autodetect.params.TimeRange;
+import org.elasticsearch.xpack.ml.job.task.JobTask;
 
 import java.io.InputStream;
 import java.util.Optional;
@@ -34,8 +35,7 @@ public class TransportPostDataAction extends TransportJobTaskAction<PostDataActi
     }
 
     @Override
-    protected void taskOperation(PostDataAction.Request request, TransportOpenJobAction.JobTask task,
-                                 ActionListener<PostDataAction.Response> listener) {
+    protected void taskOperation(PostDataAction.Request request, JobTask task, ActionListener<PostDataAction.Response> listener) {
         TimeRange timeRange = TimeRange.builder().startTime(request.getResetStart()).endTime(request.getResetEnd()).build();
         DataLoadParams params = new DataLoadParams(timeRange, Optional.ofNullable(request.getDataDescription()));
         try (InputStream contentStream = request.getContent().streamInput()) {
